requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.32.0
pandas==2.2.1
fpdf2==2.7.8
pytest==8.1.1
